backend/requirements.txt
annotated-types==0.5.0
anyio==3.7.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.6
dnspython==2.4.2
exceptiongroup==1.1.3
fastapi==0.103.2
Flask==2.3.2
Flask-Cors==4.0.0
idna==3.4
importlib-metadata==7.2.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
motor==3.3.1
numpy==1.26.0
pandas==2.1.1
pydantic==2.4.2
pydantic-core==2.10.1
pymongo==4.5.0
PyMuPDF==1.24.6
PyMuPDFb==1.24.6
pypdf2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing-extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
Werkzeug==2.3.6
zipp==3.19.2
